[CRES-80] refactor: 스터디폼을 컴포넌트로 분리 #82
